Updated the prohibited list, and fixed a type casting in hough_lines,
Added Filter2D with Border Replicate and No extra extrapolation to the prohibited list to prevent Laplacian Filter accuracy failures.

Added LUT to the prohibited list to address the LUT failures in the core accuracy test.

+++ modules/imgproc/src/opencl/hough_lines.cl
@@ -271,9 +271,9 @@ __kernel void get_lines(__global const uchar * accum_ptr, int accum_step, int ac
                 return;
 
             for (;;)
-            {
-                if (*(src_ptr + mad24(p0.y, src_step, p0.x + src_offset)))
-                {
+            {             	
+                if (*(src_ptr + mad24((int) p0.y, src_step, ((int) p0.x) + src_offset)))
+                {                
                     gap = 0;
 
                     if (!inLine)
